services/lakehouse/pyproject.toml
[project]
name = "lakehouse-api"
version = "1.0.0"
description = "Lakehouse asset storage and retrieval REST API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "motor>=3.3.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.9",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.4.0",
    "nanoid>=2.0.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio", "httpx"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
